pm-document/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.1
confluent-kafka==2.8.0
py-eureka-client==0.11.10
httpx==0.28.1
PyMuPDF==1.25.3
python-docx==1.1.2
langchain-text-splitters==0.3.5
